Silence another "possible loss of precision" warning.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/jk/trunk@1700560 13f79535-47bb-0310-9956-ffa450edef68

+++ native/common/jk_util.c
@@ -1910,7 +1910,7 @@ int jk_check_buffer_size()
             max_len = len;
         props++;
     }
-    return JK_MAX_ATTRIBUTE_NAME_LEN - max_len;
+    return JK_MAX_ATTRIBUTE_NAME_LEN - (int)max_len;
 }
 
 /*
